separated WeatherDisplay into two separate child components - MetarDisplay and TafDisplay - to better apply separation of concerns. Additionally moved all types and interfaces into a separate directory for scalability purposes, and renamed hooks folder into provider folder appropriately
