+++ front-end/src/components/user/SignUpModal.tsx
@@ -5,7 +5,7 @@ import Modal from "react-bootstrap/Modal";
 import Nav from "react-bootstrap/Nav";
 import InputGroup from "react-bootstrap/InputGroup";
 import { signupUser } from "../../utilities/userUtilities";
-import { ContextType } from "../../utilities/userUtilities";
+import { ContextType } from "../../types/userTypes";
 
 const SignUpForm: React.FC<ContextType> = ({
   user,
