+++ front-end/src/components/user/UserInfoForm.tsx
@@ -2,7 +2,8 @@ import { FormEventHandler, ReactElement, useState } from "react";
 import Form from "react-bootstrap/Form";
 import InputGroup from "react-bootstrap/InputGroup";
 import Button from "react-bootstrap/Button";
-import { ContextType, User, changeUserInfo } from "../../utilities/userUtilities";
+import { changeUserInfo } from "../../utilities/userUtilities";
+import { ContextType, User } from "../../types/userTypes";
 
 const UserInfoForm: React.FC<ContextType> = ({
   user,
