+++ front-end/src/utilities/flights/flightUtilities.tsx
@@ -1,38 +1,7 @@
 import { AxiosResponse } from "axios";
 import { api } from "../axiosConfig";
-import { Brief } from "./briefUtilities";
+import { Flight } from "../../types/flightTypes";
 // All flight crud utilities
-
-/**
- * The Flight interface defines the properties of a flight, which consists
- * of:
- * - Flight id
- * - Aircraft tail number
- * - Callsign
- * - Aircraft type and model
- * - Pilot responsible for the flight and receiving weather briefs
- * - Origin airport's icao code
- * - Destination airport's icao code
- *  - Both of which must be in the user's stored airports or there will be
- *    an error trying to post the flight
- * - The flight level of the flight
- * - The takeoff time
- * - The arrival time
- * - An array of briefs associated with the flight.
- */
-export interface Flight {
-  id: number;
-  tailNumber: number;
-  callsign: string;
-  aircraftTypeModel: string;
-  pilotResponsible: string;
-  origin: string;
-  destination: string;
-  flightLevel: number;
-  takeoffTime: string;
-  arrivalTime: string;
-  briefs: Brief[];
-}
 
 /**
  * This function takes all of the required data for a flight record and makes
